requirements.txt
asgiref==3.7.2
Django==3.2.20
pytz==2025.2
sqlparse==0.4.4
typing-extensions==4.7.1
psycopg2-binary==2.9.9
Pillow==9.2.0 # handle image
djangorestframework>=3.14.0
Faker==25.4.0
